Avoid deprecation warnings

+++ pyqtgraph/parametertree/ParameterItem.py
@@ -161,8 +161,8 @@ class ParameterItem(QtGui.QTreeWidgetItem):
         self.setText(0, self.param.title())
         fm = QtGui.QFontMetrics(self.font(0))
         size = fm.size(QtCore.Qt.TextSingleLine, self.text(0))
-        size.setHeight(size.height() * 1.35)
-        size.setWidth(size.width() * 1.15)
+        size.setHeight(int(size.height() * 1.35))
+        size.setWidth(int(size.width() * 1.15))
         self.setSizeHint(0, size)
 
     def limitsChanged(self, param, limits):
